chore: extract error fs test package
